+++ example/lib/src/storybook/storybook.dart
@@ -12,6 +12,7 @@ import 'package:example/src/storybook/stories/circular_loader.dart';
 import 'package:example/src/storybook/stories/circular_progress.dart';
 import 'package:example/src/storybook/stories/dot_indicator.dart';
 import 'package:example/src/storybook/stories/drawer.dart';
+import 'package:example/src/storybook/stories/dropdown.dart';
 import 'package:example/src/storybook/stories/icons.dart';
 import 'package:example/src/storybook/stories/linear_loader.dart';
 import 'package:example/src/storybook/stories/linear_progress.dart';
@@ -129,6 +130,7 @@ class StorybookPage extends StatelessWidget {
             CircularProgressStory(),
             DotIndicatorStory(),
             DrawerStory(),
+            DropdownStory(),
             IconsStory(),
             LinearLoaderStory(),
             LinearProgressStory(),